+++ lib/managers/ar_session_manager.dart
@@ -1,9 +1,12 @@
-import 'dart:io';
 import 'dart:typed_data';
+
+import 'package:ar_flutter_plugin/datatypes/config_planedetection.dart';
+import 'package:ar_flutter_plugin/models/ar_anchor.dart';
 import 'package:ar_flutter_plugin/models/ar_hittest_result.dart';
+import 'package:ar_flutter_plugin/utils/json_converters.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:ar_flutter_plugin/datatypes/config_planedetection.dart';
+import 'package:vector_math/vector_math_64.dart';
 
 // Type definitions to enforce a consistent use of the API
 typedef ARHitResultHandler = void Function(List<ARHitTestResult> hits);
@@ -31,6 +34,68 @@ class ARSessionManager {
     if (debug) {
       print("ARSessionManager initialized");
     }
+  }
+
+  /// Returns the camera pose in Matrix4 format with respect to the world coordinate system of the [ARView]
+  Future<Matrix4?> getCameraPose() async {
+    try {
+      final serializedCameraPose = await _channel.invokeMethod<List<dynamic>>('getCameraPose', {});
+      return MatrixConverter().fromJson(serializedCameraPose!);
+    } catch (e) {
+      print('Error caught: ' + e.toString());
+      return null;
+    }
+  }
+
+  /// Returns the given anchor pose in Matrix4 format with respect to the world coordinate system of the [ARView]
+  Future<Matrix4?> getPose(ARAnchor anchor) async {
+    try {
+      if (anchor.name.isEmpty) {
+        throw Exception("Anchor can not be resolved. Anchor name is empty.");
+      }
+      final serializedCameraPose = await _channel.invokeMethod<List<dynamic>>('getAnchorPose', {
+        "anchorId": anchor.name,
+      });
+      return MatrixConverter().fromJson(serializedCameraPose!);
+    } catch (e) {
+      print('Error caught: ' + e.toString());
+      return null;
+    }
+  }
+
+  /// Returns the distance in meters between @anchor1 and @anchor2.
+  Future<double?> getDistanceBetweenAnchors(ARAnchor anchor1, ARAnchor anchor2) async {
+    var anchor1Pose = await getPose(anchor1);
+    var anchor2Pose = await getPose(anchor2);
+    var anchor1Translation = anchor1Pose?.getTranslation();
+    var anchor2Translation = anchor2Pose?.getTranslation();
+    if (anchor1Translation != null && anchor2Translation != null) {
+      return getDistanceBetweenVectors(anchor1Translation, anchor2Translation);
+    } else {
+      return null;
+    }
+  }
+
+  /// Returns the distance in meters between @anchor and device's camera.
+  Future<double?> getDistanceFromAnchor(ARAnchor anchor) async {
+    Matrix4? cameraPose = await getCameraPose();
+    Matrix4? anchorPose = await getPose(anchor);
+    Vector3? cameraTranslation = cameraPose?.getTranslation();
+    Vector3? anchorTranslation = anchorPose?.getTranslation();
+    if (anchorTranslation != null && cameraTranslation != null) {
+      return getDistanceBetweenVectors(anchorTranslation, cameraTranslation);
+    } else {
+      return null;
+    }
+  }
+
+  /// Returns the distance in meters between @vector1 and @vector2.
+  double getDistanceBetweenVectors(Vector3 vector1, Vector3 vector2) {
+    num dx = vector1.x - vector2.x;
+    num dy = vector1.y - vector2.y;
+    num dz = vector1.z - vector2.z;
+    double distance = sqrt(dx * dx + dy * dy + dz * dz);
+    return distance;
   }
 
   Future<void> _platformCallHandler(MethodCall call) {